+++ apps/site/lib/predicted_schedule.ex
@@ -16,6 +16,42 @@ defmodule PredictedSchedule do
           prediction: Prediction.t() | nil
         }
 
+  def get(route_id, stop_id, opts \\ []) do
+    schedules_fn = &Schedules.Repo.by_route_ids/2
+    now = Keyword.get(opts, :now)
+    direction_id = Keyword.get(opts, :direction_id)
+
+    schedules =
+      [route_id]
+      |> schedules_fn.(
+        stop_ids: stop_id,
+        min_time: now,
+        direction_id: direction_id
+      )
+
+    [route: route_id, stop: stop_id, min_time: now, direction_id: direction_id]
+    |> Predictions.Repo.all()
+    |> PredictedSchedule.group(schedules)
+    |> case do
+      [_ | _] = ps ->
+        ps
+
+      [] ->
+        # if there are no schedules left for today, get schedules for tomorrow
+        PredictedSchedule.group(
+          [],
+          schedules_fn.(
+            [route_id],
+            stop_ids: stop_id,
+            direction_id: direction_id,
+            date: Util.tomorrow_date(now)
+          )
+        )
+    end
+    |> Enum.reject(&PredictedSchedule.last_stop?/1)
+    |> Enum.reject(&(PredictedSchedule.time(&1) == nil))
+  end
+
   @doc """
   The given predictions and schedules will be merged together according to
   stop_id and trip_id to create PredictedSchedules. The final result is a sorted list of
